for UL scales and smearings

+++ RecoEgamma/EgammaTools/src/ElectronEnergyCalibrator.cc
@@ -40,8 +40,8 @@ calibrate(reco::GsfElectron &ele, unsigned int runNumber,
 	  const float smearNrSigma, 
 	  const ElectronEnergyCalibrator::EventType eventType) const
 {
-  const float scEtaAbs = std::abs(ele.superCluster()->eta());
-  const float et = ele.ecalEnergy() / cosh(scEtaAbs);
+  const double scEtaAbs = std::abs(ele.superCluster()->eta());
+  const double et = ele.ecalEnergy() / cosh(scEtaAbs);
 
   if (et < minEt_ || edm::isNotFinite(et) ) {
     std::array<float,EGEnergySysIndex::kNrSysErrs> retVal;
